* 메일   - 사원삭제 시, 이메일 계정도 삭제되도록 함.
git-svn-id: svn://svn.opensol2014.com/repo1/ezEKP4Java@3146 746f5ef6-3e3b-473a-8c3d-618c0f8083c2

+++ src/main/java/egovframework/ezEKP/ezOrgan/web/EzOrganAdminController.java
@@ -430,9 +430,19 @@ public class EzOrganAdminController extends EgovFileMngUtil{
 	@RequestMapping(value = "/admin/ezOrgan/delUser.do")
 	public void delUser(HttpServletRequest request, HttpServletResponse response) throws Exception{
 		String cn[] = request.getParameter("cn").split(",");
+		
+		// dhlee
+		String domain = config.getProperty("config.DomainName");		
+		// dhlee - end
 				
-		for (int i=0; i < cn.length; i++) {			
-			ezOrganAdminService.deleteDBData(cn[i], "user");
+		for (int i=0; i < cn.length; i++) {
+			// dhlee
+			int rc = ezEmailUserAdminService.removeUser(cn[i] + "@" + domain);
+			
+			if (rc == 0) { // 성공			
+				ezOrganAdminService.deleteDBData(cn[i], "user");
+			}
+			// dhlee - end
 		}		
 	}
